Delete of Neutron L4LB objects is idempotent
This ensures that the deletion succeeds even if the Neutron
objects or their translated MidoNet counterparts have been deleted
from ZooKeeper.

Fixes: MI-1005

+++ midonet-cluster/src/main/scala/org/midonet/cluster/services/c3po/translators/LoadBalancerPoolTranslator.scala
@@ -53,12 +53,15 @@ class LoadBalancerPoolTranslator
     }
 
     override protected def translateDelete(tx: Transaction,
-                                           npool: NeutronLoadBalancerPool)
+                                           nPool: NeutronLoadBalancerPool)
     : Unit = {
-        val pool = tx.get(classOf[Pool], npool.getId)
+        if (!tx.exists(classOf[Pool], nPool.getId)) {
+            return
+        }
+        val pool = tx.get(classOf[Pool], nPool.getId)
         val lbId = pool.getLoadBalancerId // if !hasLoadBalancerId it's a bug
         val lb = tx.get(classOf[LoadBalancer], lbId)
-        tx.delete(classOf[Pool], npool.getId, ignoresNeo = true)
+        tx.delete(classOf[Pool], nPool.getId, ignoresNeo = true)
         if (lb.getPoolIdsCount == 1) {
             tx.delete(classOf[LoadBalancer], lbId, ignoresNeo = true)
         }
